cloud_data/requirements.txt
python-dotenv
requests
psycopg2-binary
python-dateutil
numpy
pandas